+++ stiffgwpy_fast/cobaya/likelihoods/LIGO_SGWB/__init__.py
@@ -1,1 +1,6 @@
+"""LVK stochastic-background cross-correlation likelihood package.
+
+中文：导出 LVK O1/O2/O3 SGWB 交叉相关似然类。
+"""
+
 from .LVK_SGWB_CC import LVK_SGWB_CC